+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 2023/12/16
+
+- Add "Load more data" button to the XP Tracker if you want to see older data. By default only max `60` days of data is
+  stored.
+
 ## 2023/12/15
 
 - Show which skills have leveled up in the XP Tracker.